requirements.txt
cvss==2.3
PyYAML==6.0
requests==2.25.1
xmltodict==0.13.0
